@@ -1,6 +1,5 @@
-class GemInstaller
-  include SuckerPunch::Job
-  workers 16
+class GemInstallerJob < ApplicationJob
+  queue_as :default
 
   def perform(gem_name, version = nil)
     SuckerPunch.logger.info "install #{gem_name} #{version}"
@@ -9,9 +8,9 @@ class GemInstaller
       # NOTE: uninstall all versions of `gem_name` then install it for upgrade/downgrade
       pl.uninstall! if pl.installed?
       pl.install!
-      SuckerPunch.logger.info "installed #{gem_name} #{version}"
+      logger.info "installed #{gem_name} #{version}"
     rescue Plugin::GemError
-      SuckerPunch.logger.warn "installing #{gem_name} #{version} is failed"
+      logger.warn "installing #{gem_name} #{version} is failed"
     end
   end
 end
