@@ -35,7 +35,10 @@ module FluentdUi
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     config.i18n.default_locale = 'en'
-    config.autoload_paths += %W(#{config.root}/app/workers #{config.root}/lib)
+    config.i18n.available_locales = %i(en ja)
+    config.autoload_paths += %W(#{config.root}/lib)
+
+    config.active_job.queue_adapter = :sucker_punch
 
     # NOTE: currently, fluentd-ui does not using ActiveRecord, and using Time.now instead of Time.zone.now for each different TZ for users.
     #       If AR will be used, please comment in and check timezone.
