@@ -1,1 +1,2 @@
-AllPluginCheckUpdate.new.async.perform
+require "plugin" # Avoid: RuntimeError Circular dependency detected while autoloading constant Plugin
+AllPluginCheckUpdateJob.perform_later
